+++ inst/NEWS.md
@@ -1,3 +1,10 @@
+Changes in Version 1.4.1 (2024-09-23)
+========================================================
+* Fix for `toBibLaTeX` and `toBibTeX` non-ASCII name list fields. Add escape
+hatch if `tools::encoded_text_to_latex` fails to convert name lists to valid
+LaTeX (observed for Japanese names in #106 h/t @kijinosu).
+
+
 Changes in Version 1.4.0 (2022-09-30)
 ========================================================